objs in main asm (read desc)
I've done this because I do prefer it this way, yes as of right now it is quite messy and the main asm file is now quite big but that is probably because of the lack of tab usage and stuff like mappings are still unsplit.

When there is a reasonable amount of comments and when the constants get replaced (probably to s2 git, this doesn't mean this disassembly will be exactly like s2 git tho), objects will be resplit into a more organised way in the split-objects branch as I know that people do prefer it that way.